Move Chunker to separate file
