+++ .gitignore
@@ -130,6 +130,7 @@ dmypy.json
 data
 output
 benchmark
+trash
 checkpoints
 dbg
 .obsidian
