+++ ArticleReader/Narrator.py
@@ -2,6 +2,7 @@
 import torch
 import torchaudio
 from speechbrain.inference import Tacotron2, HIFIGAN
+from datetime import timedelta
 
 class Narrator:
     def __init__(self, tts_model = None, vocoder_model = None):
@@ -175,6 +176,45 @@ class Narrator:
     def save_audio(self, output_wav, waveform):
         torchaudio.save(output_wav, waveform, 22050, format="wav")
         print(f"Audio saved to {output_wav}")
+
+
+
+    def generate_srt(self, sentences, durations, output_file="output.srt"):
+        """
+        Generates an SRT file from a list of sentences and durations.
+
+        Args:
+            sentences (list of str): The sentences to be shown as subtitles.
+            durations (list of int): Durations (in seconds) for each sentence in chronological order.
+            output_file (str): Path to save the SRT file.
+
+        Returns:
+            None
+        """
+
+        def format_timestamp(seconds):
+            """Formats seconds into SRT timestamp format."""
+            td = timedelta(seconds=seconds)
+            total_seconds = int(td.total_seconds())
+            milliseconds = int(td.microseconds / 1000)
+            hours, remainder = divmod(total_seconds, 3600)
+            minutes, seconds = divmod(remainder, 60)
+            return f"{hours:02}:{minutes:02}:{seconds:02},{milliseconds:03}"
+
+        with open(output_file, "w") as file:
+            current_time = 0  # Start time in seconds
+            for i, (sentence, duration) in enumerate(zip(sentences, durations), start=1):
+                start_time = format_timestamp(current_time)
+                end_time = format_timestamp(current_time + duration)
+                current_time += duration  # Increment time by the duration of the sentence
+
+                # Write SRT entry
+                file.write(f"{i}\n")
+                file.write(f"{start_time} --> {end_time}\n")
+                file.write(f"{sentence.strip()}\n\n")
+
+        print(f"SRT file saved to {output_file}")
+
 
     def save_video(self, output_file):
         # ffmpeg -loop 1 -i image.png -i 24.12.08-17.wav -c:v libx264 -tune stillimage -c:a aac -b:a 192k -pix_fmt yuv420p -shortest output.mp4
